requirements.txt
requests == 2.31.0
jsonschema == 4.20.0
pytest == 7.4.3
selene==2.0.0rc4
allure-pytest==2.13.2